split the hash slide out of fill window
as a preparation to optimize it.

"Sliding" happens when the compression window/2 gets moved.
In this case the hash values are adapted, which means
basically an unsigned subtration with saturation (no wrap
at zero).

Many processor have an instruction for this, but the
compiler, trapped in the C-type systems, is seldomly
capable to exploit this.