add PPC versions of slhash
First an Altivec version which gives a nice speedup on
"time minigzip $LARGEFILE":

============ orig ===============
32.50 real
33.55 real
33.51 real
========== altivec ==============
29.12 real
29.29 real
28.88 real
a ~12.5% speedup

But there is also room for improvement on chips whithout
Altivec.
PPC is missing a conditional move until recently (Power
ISA v2.06, a.k.a POWER7). On the other hand PPC is known
to not cope well with conditional branches. So with a little
trick one can force creation of a pseudo conditional move.

again "time minigzip $LARGEFILE":
============ orig ===============
32.88 real
33.34 real
32.99 real
============= ppc ===============
31.36 real
31.12 real
31.52 real
~5.6% speedup