add bfin version of slhash
The code is 2.7 times faster in cycle count.
The total speedup for "time minigzip $LARGEFILE" is a
little unstable between ~3% and ~8% (using not sooo large
$LARGEFILE since RAMFS).

One problem for blackfin is it only has signed saturation,
no unsigned saturation. With a little trickery this is usable,
but reduces the value range usable with the faster
instructions, so the 64k window size cannot use the faster
code.